requirements.txt
pdfplumber
google-generativeai
openpyxl
rich
click
